+++ paperback-64e9d9ddf9ec4a5d1cbdb3a14095f5e3da8bc67e/utilities/misc_functions.lua
@@ -351,29 +351,24 @@ function PB_UTIL.add_tag(tag, event)
   end
 end
 
----Gets a pseudorandom consumable from the Consumables pool (Soul and Black Hole included)
----@param seed string
----@param soulable boolean
----@return table
-function PB_UTIL.poll_consumable(seed, soulable)
+--- Gets a random consumable type
+--- @param seed string
+--- @return SMODS.ConsumableType
+function PB_UTIL.poll_consumable_type(seed)
   local types = {}
 
-  for k, v in pairs(SMODS.ConsumableTypes) do
-    types[#types + 1] = k
+  for _, v in pairs(SMODS.ConsumableTypes) do
+    types[#types + 1] = v
   end
 
-  return SMODS.create_card {
-    set = pseudorandom_element(types, pseudoseed(seed)),
-    area = G.consumables,
-    soulable = soulable,
-    key_append = seed,
-  }
+  return pseudorandom_element(types, pseudoseed(seed))
 end
 
 --- Tries to spawn a card into either the Jokers or Consumeable card areas, ensuring
 --- that there is space available.
 --- DOES NOT TAKE INTO ACCOUNT ANY OTHER AREAS
 --- @param args CreateCard | { instant: boolean?, func: function? } same arguments passed to SMODS.create_card, with the addition of 'instant' and 'func'
+--- @return boolean? spawned whether the card was able to spawn
 function PB_UTIL.try_spawn_card(args)
   local is_joker = (args.set == 'Joker' or args.key and args.key:sub(1, 1) == 'j')
   local area = args.area or (is_joker and G.jokers) or G.consumeables
@@ -397,6 +392,8 @@ function PB_UTIL.try_spawn_card(args)
     if args.func and type(args.func) == "function" then
       args.func()
     end
+
+    return true
   end
 end
 
@@ -481,7 +478,7 @@ end
 
 ---This function is basically a copy of how the base game does the flipping animation
 ---on playing cards when using a consumable that modifies them
----@param card table
+---@param card table?
 ---@param cards_to_flip table?
 ---@param action function?
 ---@param sound string?
@@ -491,15 +488,17 @@ function PB_UTIL.use_consumable_animation(card, cards_to_flip, action, sound)
     cards_to_flip = { cards_to_flip }
   end
 
-  G.E_MANAGER:add_event(Event {
-    trigger = 'after',
-    delay = 0.4,
-    func = function()
-      play_sound(sound or 'tarot1')
-      card:juice_up(0.3, 0.5)
-      return true
-    end
-  })
+  if card then
+    G.E_MANAGER:add_event(Event {
+      trigger = 'after',
+      delay = 0.4,
+      func = function()
+        play_sound(sound or 'tarot1')
+        card:juice_up(0.3, 0.5)
+        return true
+      end
+    })
+  end
 
   if cards_to_flip then
     for i = 1, #cards_to_flip do
@@ -623,6 +622,17 @@ function PB_UTIL.get_lowest_hand_discard()
     return { amt = hands, hands = true }
   else
     return { amt = discards, discards = true }
+  end
+end
+
+--- Returns the key of the Planet card for the specified poker hand
+--- @param hand_name string the name of the poker hand, like "Four of a Kind"
+--- @return string?
+function PB_UTIL.get_planet_for_hand(hand_name)
+  for _, v in ipairs(G.P_CENTER_POOLS.Planet) do
+    if v.config and v.config.hand_type == hand_name then
+      return v.key
+    end
   end
 end